Adjusting the input and output card

+++ educarte/lib/Interector/models/entry_and_exit_modal.dart
@@ -57,7 +57,7 @@ class AccessControls {
 
 class ContractedHour {
   String? id;
-  int? hours;
+  double? hours;
   int? status;
   String? startDate;
   String? endDate;
